Remove obsolete bot and tests, add updated BOT version
Both the MyBot and MyBotTests files were deleted because they became obsolete and did not meet the required performance.

Brought in an updated version of the 'MyBot' and 'MyBotTests'scripts. The updated bot is more optimized, providing improved performance. The bot utilizes a transposition table enhancing the bot's ability to evaluate the board positions. Additional improvements include a smart ordering of moves and checks, and efficient pruning.

Also, adjusted the max token count from 1024 to 2048 in 'Settings.cs'. This adjustment is expected to help cater to the increased computational needs of the updated bot.

As part of this commit, we also created a new file, stockfish-output.txt, for storing the output from the Stockfish engine for better debugging and profile analysis.