"Integrate SearchAllCaptures method into bot logic
The main logic of the QuickMateBot has been changed to make use of the new SearchAllCaptures function, to better evaluate potential capture moves. This function improves the bot's decision-making when it comes to making capture moves, ensuring it considers the potential value of a capture move before deciding to execute it. This should lead to more sophisticated